+++ src/main/scala/sc4pac/Find.scala
@@ -48,7 +48,7 @@ object Find {
           val repoUris = context.repositories.map(_.baseUri)
           context.logger.log(s"Could not find metadata of ${module}. Trying to update channel contents.")
           for {
-            repos   <- Sc4pac.initializeRepositories(repoUris, context.cache, channelContentsTtl = Some(60.seconds))
+            repos   <- Sc4pac.initializeRepositories(repoUris, context.cache, Constants.channelContentsTtlShort)  // 60 seconds
                         .provideLayer(zio.ZLayer.succeed(ScopeRoot(context.scopeRoot)))
             result  <- tryAllRepos(repos, context)  // 2nd try
           } yield result